Use Altered Directory Structure to Store Name Files
To build its search database, you must run the `generate-names.pl` script which
indexes the names that will become searchable. Names were previously stored in
a large `names.txt` file, but this change makes it easier to separate out these
names by track in a way that is acceptable to the original perl script. In the
future, this will hopefully allow a more useful feature description in the search
window.